+++ custom_crm/jaybutani_crm_customisation/__manifest__.py
@@ -1,7 +1,7 @@
 # -*- coding: utf-8 -*-
 {
     'name': 'CRM Customisation',
-    'version': '18.0.0.0.0',
+    'version': '18.0.1.0.0',
     'author': 'Jay Butani',
     'sequence': 5,
     'summary': 'CRM Customisation Module',
@@ -11,9 +11,10 @@
         'views/sale_order_views.xml',
         'views/stock_picking_views.xml',
         'views/mrp_production_views.xml',
-        'views/stock_move_views.xml',
+        'views/project_project_views.xml',
+        'views/account_move_views.xml'
     ],
-    'depends':['sale_management','contacts','stock','crm', 'purchase','mrp', 'project'],
+    'depends':['sale_management','contacts','stock','crm','mrp', 'project', 'sale_project','account'],
     'application': True,
     'license': 'LGPL-3',
 }