+++ custom_crm/jaybutani_crm_customisation/models/__init__.py
@@ -5,3 +5,6 @@ from . import stock_picking
 from . import stock_rule
 from . import stock_move
 from . import mrp_production
+from . import project_project
+from . import sale_order_line
+from . import account_move