+++ custom_crm/jaybutani_crm_customisation/models/account_move.py
@@ -3,7 +3,7 @@
 from odoo import models, fields
 
 
-class StockMoveLine(models.Model):
-    _inherit = 'stock.move'
+class AccountMove(models.Model):
+    _inherit = 'account.move'
 
     job_name = fields.Char(string='Job Name')