+++ custom_crm/jaybutani_crm_customisation/models/stock_picking.py
@@ -6,4 +6,4 @@ from odoo import models, fields
 class StockPicking(models.Model):
     _inherit = 'stock.picking'
 
-    job_name = fields.Char(related='move_ids.job_name',string='Job Name',readonly=False)
+    job_name = fields.Char(related='move_ids.job_name', string='Job Name')