+++ custom_addons/ak_library_management/changelog.md
@@ -102,3 +102,4 @@
 
 - add one condition on template
 - add docstring on controller
+- change main.py and add sudo search